`sidepanel` moved under browser `ui/views/` subpath
This is merely a path change, which does affect a few inclusion paths
and patches.

Chromium change:
https://chromium.googlesource.com/chromium/src/+/f4096571c202d7aa7e5c9d9b8dc9edebbbeb631e

commit f4096571c202d7aa7e5c9d9b8dc9edebbbeb631e
Author: Erik Chen <erikchen@chromium.org>
Date:   Fri Jun 7 05:33:08 2024 +0000

    Move sidepanel C++ code into a single directory.

    This CL is a refactor with no intended behavior change.

    This CL was generated with 4 steps:

    (1) deleting redundant OWNER files

    (2) moving all files from chrome/browser/ui/side_panel to
    chrome/browser/ui/views/side_panel and updating BUILD.gn files.

    (3) Updating all includes with the command:
    git grep -l 'chrome\/browser\/ui\/side_panel' | xargs sed -i \
    's/chrome\/browser\/ui\/side_panel/chrome\/browser\/ui\/views\/side_panel/g'

    (4) Updating all header preprocessor macros with the command:
    git grep -l 'CHROME_BROWSER_UI_SIDE_PANEL' | xargs sed -i \
    's/CHROME_BROWSER_UI_SIDE_PANEL/CHROME_BROWSER_UI_VIEWS_SIDE_PANEL/g'

    Bug: 345499793
