+++ browser/ui/webui/speedreader/speedreader_toolbar_data_handler_impl.cc
@@ -30,7 +30,7 @@
 
 #if BUILDFLAG(ENABLE_AI_CHAT)
 #include "brave/components/ai_chat/core/browser/utils.h"
-#include "chrome/browser/ui/side_panel/side_panel_ui.h"
+#include "chrome/browser/ui/views/side_panel/side_panel_ui.h"
 #endif
 
 namespace {
